Support config commands
`git-issue` will now execute shell commands in config entries which
begin with `!` following the convention for executing shell commands in
Git aliases. This allows service API keys to be stored in any form the
user desired, e.g. a file encrypted with GPG.

The following `.gitconfig` entry will decrypt the file and use the
string printed to `stdout` for the API token to authenticate with the
GitHub API.

```gitconfig
[issue "GitHub"]
	token = !gpg --decrypt github-token.gpg
```